git-p4: Handle a missing 'fileSize' without throwing an exception.
The old code would throw KeyError and die if `self.stream_file`
contained no `'fileSize'` key, which I've observed in practice with
my employer's Perforce repo (although I have not investigated *why*
it's absent).

The new code checks for the existence of the key before trying to
use it. Similar checks are already present on lines 2668 and 2712.