git-p4: Add a `git_build_cmd` helper function. NFC.
This allows me to funnel all git commands through a single point.
My purpose is to be able to print out the commands that git-p4 is
running, so that I can learn about plumbing. However, this patch
could theoretically also be the starting point for the ability
to customize the path of `git` (e.g. to use a `git` binary that
is not the first one in your PATH).

No functional change intended.